+++ kotest-core/src/commonMain/kotlin/io/kotest/core/spec/style/behaviorSpec.kt
@@ -1,9 +1,11 @@
 package io.kotest.core.spec.style
 
+import io.kotest.core.config.Project
 import io.kotest.core.factory.TestFactory
 import io.kotest.core.factory.TestFactoryConfiguration
 import io.kotest.core.factory.build
 import io.kotest.core.spec.SpecConfiguration
+import io.kotest.core.test.TestCaseConfig
 
 /**
  * Creates a [TestFactory] from the given block.
@@ -18,10 +20,14 @@ fun behaviorSpec(block: BehaviorSpecTestFactoryConfiguration.() -> Unit): TestFa
 }
 
 class BehaviorSpecTestFactoryConfiguration : TestFactoryConfiguration(), BehaviorSpecDsl {
+   override fun defaultConfig(): TestCaseConfig = defaultTestCaseConfig ?: Project.testCaseConfig()
    override val addTest = ::addDynamicTest
 }
 
 abstract class BehaviorSpec(body: BehaviorSpec.() -> Unit = {}) : SpecConfiguration(), BehaviorSpecDsl {
+   override fun defaultConfig(): TestCaseConfig =
+      defaultTestCaseConfig ?: defaultTestCaseConfig() ?: Project.testCaseConfig()
+
    override val addTest = ::addRootTestCase
 
    init {