+++ kotest-core/src/commonMain/kotlin/io/kotest/core/spec/style/shouldSpec.kt
@@ -1,10 +1,12 @@
 package io.kotest.core.spec.style
 
 import io.kotest.Matcher
+import io.kotest.core.config.Project
 import io.kotest.core.factory.TestFactory
 import io.kotest.core.factory.TestFactoryConfiguration
 import io.kotest.core.factory.build
 import io.kotest.core.spec.SpecConfiguration
+import io.kotest.core.test.TestCaseConfig
 import io.kotest.should as shouldBeMatcher
 
 /**
@@ -36,10 +38,14 @@ fun shouldSpec(block: ShouldSpecTestFactoryConfiguration.() -> Unit): TestFactor
 }
 
 class ShouldSpecTestFactoryConfiguration : TestFactoryConfiguration(), ShouldSpecDsl {
+   override fun defaultConfig(): TestCaseConfig = defaultTestCaseConfig ?: Project.testCaseConfig()
    override val addTest = ::addDynamicTest
 }
 
 abstract class ShouldSpec(body: ShouldSpec.() -> Unit = {}) : SpecConfiguration(), ShouldSpecDsl {
+   override fun defaultConfig(): TestCaseConfig =
+      defaultTestCaseConfig ?: defaultTestCaseConfig() ?: Project.testCaseConfig()
+
    override val addTest = ::addRootTestCase
 
    init {