+++ kotest-core/src/commonMain/kotlin/io/kotest/core/spec/style/stringSpec.kt
@@ -1,9 +1,11 @@
 package io.kotest.core.spec.style
 
+import io.kotest.core.config.Project
 import io.kotest.core.factory.TestFactory
 import io.kotest.core.factory.TestFactoryConfiguration
 import io.kotest.core.factory.build
 import io.kotest.core.spec.SpecConfiguration
+import io.kotest.core.test.TestCaseConfig
 
 /**
  * Creates a [TestFactory] from the given block.
@@ -21,10 +23,14 @@ fun stringSpec(block: StringSpecTestFactoryConfiguration.() -> Unit): TestFactor
  * Decorates a [TestFactoryConfiguration] with the StringSpec DSL.
  */
 class StringSpecTestFactoryConfiguration : TestFactoryConfiguration(), StringSpecDsl {
+   override fun defaultConfig(): TestCaseConfig = defaultTestCaseConfig ?: Project.testCaseConfig()
    override val addTest = ::addDynamicTest
 }
 
 abstract class StringSpec(body: StringSpec.() -> Unit = {}) : SpecConfiguration(), StringSpecDsl {
+   override fun defaultConfig(): TestCaseConfig =
+      defaultTestCaseConfig ?: defaultTestCaseConfig() ?: Project.testCaseConfig()
+
    override val addTest = ::addRootTestCase
 
    init {